scraper/requirements.txt
httpx==0.28.1
xmltodict==0.14.2
python-dotenv==1.0.0
supabase==2.9.0
websockets==15.0.1
